+++ ui/components/ui/content-layout/content-layout.tsx
@@ -1,9 +1,8 @@
-import { ReactNode, Suspense, use } from "react";
+"use client";
 
-import { getUserInfo } from "@/actions/users/users";
+import { ReactNode } from "react";
 
 import { Navbar } from "../nav-bar/navbar";
-import { SkeletonContentLayout } from "./skeleton-content-layout";
 
 interface ContentLayoutProps {
   title: string;
@@ -12,13 +11,9 @@ interface ContentLayoutProps {
 }
 
 export function ContentLayout({ title, icon, children }: ContentLayoutProps) {
-  const user = use(getUserInfo());
-
   return (
     <>
-      <Suspense fallback={<SkeletonContentLayout />}>
-        <Navbar title={title} icon={icon} user={user} />
-      </Suspense>
+      <Navbar title={title} icon={icon} />
       <div className="px-6 py-4 sm:px-8 xl:px-10">{children}</div>
     </>
   );