+++ ui/components/ui/custom/index.ts
@@ -8,4 +8,5 @@ export * from "./custom-loader";
 export * from "./custom-radio";
 export * from "./custom-section";
 export * from "./custom-server-input";
+export * from "./custom-table-link";
 export * from "./custom-textarea";